+++ sdk/disconnectedoperations/arm-disconnectedoperations/src/models/index.ts
@@ -35,7 +35,7 @@ export {
   ErrorAdditionalInfo,
   DisconnectedOperationCreateOrUpdate,
   DisconnectedOperationPropertiesCreateOrUpdate,
-  BillingConfigurationCreateCreateOrUpdate,
+  BillingConfigurationCreateOrUpdate,
   DisconnectedOperationUpdate,
   DisconnectedOperationUpdateProperties,
   DisconnectedOperationDeploymentManifest,